feat(updated): add useProfile & useProfileDetails & useRestorePassword hooks
