+++ src/screens/Profile/hooks/index.tsx
@@ -1,0 +1,1 @@
+export {useProfile} from './useProfile'
