+++ src/screens/ProfileDetails/hooks/index.tsx
@@ -1,0 +1,1 @@
+export {useProfileDetails} from './useProfileDetails'
