+++ src/screens/RestorePassword/hooks/index.tsx
@@ -1,0 +1,1 @@
+export {useRestorePassword} from './useRestorePassword'
